requirements.txt
crewai==0.148.0
langchain==0.3.26
langchain-community==0.3.27
openai==1.97.0
ipykernel==6.30.0
jupyter==1.1.1
python-dotenv==1.1.1
pdfkit==1.0.0
markdown==3.8.2